wip: refactor code
